cleaned a lot of functions

+++ ft_strmapi.c
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	char				*str;
-	unsigned int		i;
+	char	*str;
+	size_t	i;
 
 	if (!s || !f)
 		return (NULL);
